login utility work in progress

+++ AdministrativeDivisionOfPoland-Backend/src/main/java/pl/edu/pwr/administrativedivisionofpoland/administrativedivisionofpolandbackend/Repositories/VoivodeshipRepository.java
@@ -4,11 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pl.edu.pwr.administrativedivisionofpoland.administrativedivisionofpolandbackend.Entities.Voivodeship;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface VoivodeshipRepository extends JpaRepository<Voivodeship, Long> {
+public interface VoivodeshipRepository extends JpaRepository<Voivodeship, Integer> {
 
     Optional<Voivodeship> findById(int id);
 }
